+++ sdk/keyvault/azure-security-keyvault-certificates/src/test/java/com/azure/security/keyvault/certificates/CertificateClientTest.java
@@ -862,7 +862,7 @@ public class CertificateClientTest extends CertificateClientTestBase {
                 certificateClient.importCertificate(importCertificateOptions);
 
             assertTrue("db1497bc2c82b365c5c7c73f611513ee117790a9"
-                .equalsIgnoreCase(importedCertificate.getProperties().getX509ThumbprintString()));
+                .equalsIgnoreCase(importedCertificate.getProperties().getX509ThumbprintAsString()));
             assertEquals(importCertificateOptions.isEnabled(), importedCertificate.getProperties().isEnabled());
 
             // Load the CER part into X509Certificate object