+++ sdk/keyvault/azure-security-keyvault-certificates/CHANGELOG.md
@@ -5,7 +5,7 @@ Changes when compared to the last stable release (`4.5.9`) include:
 
 ### Features Added
 - Added support for service version `7.5`.
-- Added `CertificateProperties.getX509ThumbprintString()` to return the hexadecimal string representation of the SHA-1 hash of a certificate.
+- Added `CertificateProperties.getX509ThumbprintAsString()` to return the hexadecimal string representation of the SHA-1 hash of a certificate.
 
 ### Other Changes